ml-service/requirements.txt
transformers>=4.30.0
datasets>=2.12.0
torch>=2.0.0
scikit-learn>=1.3.0
accelerate>=0.20.0
flask>=2.3.0
flask-cors>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pymupdf>=1.22.0
werkzeug>=2.3.0
tensorboard>=2.13.0